+++ node/npm/glide/index.ts
@@ -191,6 +191,8 @@ function initialize() {
         ReturnTypeAttribute,
         ReturnTypeJson,
         UniversalReturnTypeJson,
+        Score,
+        ElementAndScore,
     } = nativeBinding;
 
     module.exports = {
@@ -340,6 +342,8 @@ function initialize() {
         ReturnTypeAttribute,
         ReturnTypeJson,
         UniversalReturnTypeJson,
+        Score,
+        ElementAndScore,
     };
 
     globalObject = Object.assign(global, nativeBinding);